frontend/requirements.txt
Flask
requests
Authlib
gunicorn
kafka-python